Dynamically inject current public IP as allowed
We do not want to block ourselves everytime we run terraform
when working form a IP that was added manually is not in the
password store (e.g. when in a coffee wifi, or a VPN).

Instead, we get the current IP from http://ifconfig.co and
inject it as allowed IP.