Add module to configure sub accounts
We add a module similar to the root account that would configure:

 - The account alias
 - The policies:
   - RestrictToWhitelistedIPs to avoid usage outside of
     the whitelisted ips
   - AssumeRoleTrustPolicySubAccountWithMFA that would allow
     assume role using MFA from the root account

 - The roles:
   - admin: for admin usage
   - dev: for normal task as poweuser.

The roles require use the MFA token.

We configure the provider with assume role[1], to configure the
resources in the right account. In the first run, we need to use
a different role that is created in the bootstrap of the account
instead of 'admin', because 'admin' role does not exist. Because
that we allow pass the variable `terraform_assume_role`

We do not want to create the 'admin' role outside of terraform
during bootstrap because if we do, it is difficult to import it back.
It is easier to create a different one and delete it later.

The assume role would have problems with the MFA token, which
we explain in the next commit.

[1] https://www.terraform.io/docs/providers/aws/#assume-role