Attempt to configure assume role in terraform
The current policy to assume the roles requires
to have STS credentials with MultiFactorAuthPresent for the current
credentials.

But in the current implementation you must run terraform using
credentials with the role `admin`, to be able to access the bucket
with the state and the operations in the root account.

When terraform tries to assume the role for the submodule, it does it
without asking for the MFA credentials, so assume role
fails.

The solution would be to execute the project with the root user
credentials, and then terraform assumes role for but using MFA:
 - Getting the state for S3, using assume role[1]
 - Configure root account, users

The problem is that MFA support is not implemented in terraform[2].
It would require changes in the code [3] to add logic to ask for
the MFA token:

   assumeRoleProvider.SerialNumber = aws.String("arn:awss:iam::246505230303:mfa/hector.rivas+aws.admin@keytwine.com")
   assumeRoleProviderleProvider.TokenProvider = stscreds.StdinTokenProvider

Leaving the code here for reference

[1] https://www.terraform.io/docs/backends/types/s3.html#role_arn
[2] https://github.com/hashicorp/terraform/pull/11734
[3] https://github.com/terraform-providers/terraform-provider-aws/blob/master/aws/auth_helpers.go#L211-L227