Use specific terraform-admin role without MFA
We want to be able to run terraform and assume the role without
a MFA token, as terraform does not support MFA (See previous commit).

But we still want the users to require MFA token to be able to
assume any role.

As workaround, we allow the Principal role "admin" to assume any
other role without the MFA token by adding an additional entry
in the statements of the assume role policy.