Add ALLOW policy for assume role in gfl subaccount
This is special case. Some users can be admin or assume
any other role in the subaccount gfl.

As it is a special case, we add for the time being the specific
policy, group and membership in the subaccount definition file.

In the future we can refactor this to provide a more generic
membership configuration.