Automate update of whitelisted IPs
We want to easily add the current IPs to the policy that blocks
AWS API access to whitelisted IPs, even from an IP that is blocked.

Allow anything to modify a policy is risky, as we cannot control
what is added to the policy. Because that we want to ensure
that such operation requires at least MFA, and that minimises
the possibility of access the account with leaked API creds.

For this we would create a special role `whitelist_ips_updater`,
which can:

 - update the RestrictToWhitelistedIPs policy, to add new IPs.
 - be assumed from any user with MFA token. Note that only admin
   can explicitly assume it anyway.
 - be assumed from the role `whitelist_ips_updater`. This is to
   allow change the policy in multiple accounts but with only
   input the MFA once.

The RestrictToWhitelistedIPs would now block all the operations,
except assume the role `whitelist_ips_updater`.

We move all the policies, common for root and sub accounts to a
dedicated submodule.

To actually add the IP we wrote a script that would assume
a special role using MFA token, and then iterate across each
account to read the policy & inject the current IP
retrieved from http://ifconfig.co.

One does not actually update the policy, but create a new default
version. whenever we add a new policy version, we delete the oldest
non default version to be sure we do not add more than the allowed
versions.

To use the script, you must load the user credentials and run it.
For example, using my `awssts` tool[1]:

   awssts user:hector.rivas+admin@keytwine \
      ./scripts/whitelist_current_ip.sh

[1] https://github.com/keymon/aws_key_management/blob/master/awssts.sh